Refactor page into components
